remove tools folder entirely
