+++ cmd/iiif-process-and-tile/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+       "log"
+       
+)
+
+func main() {
+
+     log.Fatal("Not implemented")
+    }
